# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Digital Courses API",
        "description": "API REST da plataforma de cursos online AulasViolão. Autenticação via JWT Bearer Token.",
        "contact": {
            "name": "Big Space Creative",
            "email": "admin@digitalcourses.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor de desenvolvimento"
        },
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/api/v1/courses": {
            "get": {
                "tags": [
                    "Cursos"
                ],
                "summary": "Listar cursos",
                "description": "Retorna os cursos visiveis para o usuario autenticado com modulos e aulas (sem video_url).",
                "operationId": "courseIndex",
                "responses": {
                    "200": {
                        "description": "Cursos listados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cursos listados com sucesso"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Cursos"
                ],
                "summary": "Criar curso",
                "description": "Cria um novo curso. Requer autenticacao de admin. Aceita thumbnail por URL ou upload.",
                "operationId": "courseStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Violao do Zero"
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    },
                                    "thumbnail": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "thumbnail_file": {
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "is_published": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Curso criado"
                    },
                    "401": {
                        "description": "Nao autenticado"
                    },
                    "422": {
                        "description": "Erro de validacao"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/courses/{course}": {
            "get": {
                "tags": [
                    "Cursos"
                ],
                "summary": "Exibir curso",
                "description": "Retorna um curso especifico com seus modulos e aulas. Cursos nao publicados ficam restritos a admin/instructor.",
                "operationId": "courseShow",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "ID do curso",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Curso encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Curso encontrado"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Curso nao encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Cursos"
                ],
                "summary": "Atualizar curso",
                "description": "Atualiza dados de um curso existente. Requer autenticacao de admin.",
                "operationId": "courseUpdate",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "ID do curso",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Curso atualizado"
                    },
                    "401": {
                        "description": "Nao autenticado"
                    },
                    "404": {
                        "description": "Curso nao encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Cursos"
                ],
                "summary": "Excluir curso",
                "description": "Remove permanentemente um curso. Requer autenticacao (admin).",
                "operationId": "courseDestroy",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "ID do curso",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Curso excluido"
                    },
                    "401": {
                        "description": "Nao autenticado"
                    },
                    "404": {
                        "description": "Curso nao encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Insira o token JWT obtido no login: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}